+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 * Update Complex lookups to use the new Search feature. Refs [UILD-686].
 * Add Hub tab to landing page. Refs [UILD-687].
 * Add custom workspace profile settings application. Refs [UILD-684].
+* Add Hub search behavior for LoC source. Refs [UILD-688].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -124,6 +125,7 @@
 [UILD-686]:https://folio-org.atlassian.net/browse/UILD-686
 [UILD-687]:https://folio-org.atlassian.net/browse/UILD-687
 [UILD-684]:https://folio-org.atlassian.net/browse/UILD-684
+[UILD-688]:https://folio-org.atlassian.net/browse/UILD-688
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].